use bfloat16

+++ main_training.py
@@ -61,9 +61,9 @@ def main(**kwargs):
         # get hf model
         model_name_or_path = cfg.model_name_or_path
         if rank == 0:
-            model = AutoModelForCausalLM.from_pretrained(model_name_or_path, low_cpu_mem_usage=True) # requires Accelerate: "pip install accelerate"
+            model = AutoModelForCausalLM.from_pretrained(model_name_or_path, low_cpu_mem_usage=True, torch_dtype=torch.bfloat16) # requires Accelerate: "pip install accelerate"
         else:
-            model_config = AutoConfig.from_pretrained(model_name_or_path)
+            model_config = AutoConfig.from_pretrained(model_name_or_path, torch_dtype=torch.bfloat16)
             with torch.device("meta"):
                 model = AutoModelForCausalLM.from_config(model_config)
